+++ changelogs/CHANGELOG.md
@@ -12,6 +12,7 @@
 -   Add Randomize Glyph to Death Tomes [\#785](https://github.com/EnigmaticaModpacks/Enigmatica9/issues/785)
 -   Dissolution Chamber can now bottle any XP fluid [\#788](https://github.com/EnigmaticaModpacks/Enigmatica9/issues/788)
 -   Cable Anchors may now be made in a sawmill at reduced material cost [\#789](https://github.com/EnigmaticaModpacks/Enigmatica9/issues/789)
+-   Bookshelves may be processed in a saw mill to extract books [\#791](https://github.com/EnigmaticaModpacks/Enigmatica9/issues/791)
 -   [Expert] Awkward potions are no longer brewable with Netherwart to promote the use of the far easier to obtain Abjuration Essence [\#751](https://github.com/EnigmaticaModpacks/Enigmatica9/issues/751)
 -   [Expert] Nether Wart may now be used to make Ethanol more efficiently [\#751](https://github.com/EnigmaticaModpacks/Enigmatica9/issues/751)
 -   [Expert] The Break Glyph no longer requires Iron, making it accessible closer to when a Tier 2 book is available [\#751](https://github.com/EnigmaticaModpacks/Enigmatica9/issues/751)
@@ -35,6 +36,7 @@
 -   [Expert] Chapter 1 quest clarifications to help guide players to things like the Storage Lectern [\#787](https://github.com/EnigmaticaModpacks/Enigmatica9/issues/787)
 -   [Expert] Various Twilight bosses now drop some essences to give the player a jump start before they get their production going [\#787](https://github.com/EnigmaticaModpacks/Enigmatica9/issues/787)
 -   [Expert] Compressed Stone may now be found in the Troll vaults [\#787](https://github.com/EnigmaticaModpacks/Enigmatica9/issues/787)
+-   [Expert] Dimensional Storage Crystals may now be produced in a Thermal saw [\#791](https://github.com/EnigmaticaModpacks/Enigmatica9/issues/791)
 
 ### 🐛 Fixed Bugs